requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic[email]==2.10.4
pydantic-settings==2.7.1

# Database
sqlalchemy==2.0.36

# Authentication
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.20

# Rate Limiting
slowapi==0.1.9

# Logging
python-json-logger==3.2.1

# Testing
pytest==8.3.4
httpx==0.28.1
pytest-cov==6.0.0

# Linting (optional)
ruff==0.8.6
